pyproject.toml
[project]
name = "elt-e-commerce"
version = "1.0.0"
description = "Comprehensive ELT pipeline for e-commerce data using Airflow, PostgreSQL, BigQuery and dbt."
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "apache-airflow>=3.1.5",
    "apache-airflow-providers-google>=19.4.0",
    "apache-airflow-providers-postgres>=6.5.1",
    "astronomer-cosmos>=1.13.1",
    "pandas>=2.3.3",
    "pytest>=9.0.2",
    "ruff>=0.14.10",
]

[tool.setuptools]
py-modules = []

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
pythonpath = "."
testpaths = [
    "./airflow/tests"
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]  # E: pycodestyle, F: pyflakes, I: isort (import sorting)
ignore = ["E501"]         # Ignore line length errors (optional, e.g. if Black handles this)
